refactor: move to mongosh-repl and use a hotfixed array
